feature/MTCE-235-Monorepo-v6-refactoring-EasyRepository
