+++ composer.json
@@ -117,6 +117,8 @@
             "EonX\\EasyPipeline\\": "packages/EasyPipeline/src",
             "EonX\\EasyRandom\\": "packages/EasyRandom/src",
             "EonX\\EasyRepository\\": "packages/EasyRepository/src",
+            "EonX\\EasyRepository\\Bundle\\": "packages/EasyRepository/bundle",
+            "EonX\\EasyRepository\\Laravel\\": "packages/EasyRepository/laravel",
             "EonX\\EasyRequestId\\": "packages/EasyRequestId/src",
             "EonX\\EasySchedule\\": "packages/EasySchedule/src",
             "EonX\\EasySecurity\\": "packages/EasySecurity/src",
@@ -163,6 +165,8 @@
             "EonX\\EasyPipeline\\Tests\\": "packages/EasyPipeline/tests",
             "EonX\\EasyRandom\\Tests\\": "packages/EasyRandom/tests",
             "EonX\\EasyRepository\\Tests\\": "packages/EasyRepository/tests",
+            "EonX\\EasyRepository\\Tests\\Unit\\": "packages/EasyRepository/tests/Unit/src",
+            "EonX\\EasyRepository\\Tests\\Unit\\Laravel\\": "packages/EasyRepository/tests/Unit/laravel",
             "EonX\\EasyRequestId\\Tests\\": "packages/EasyRequestId/tests",
             "EonX\\EasySchedule\\Tests\\": "packages/EasySchedule/tests",
             "EonX\\EasySecurity\\Tests\\": "packages/EasySecurity/tests",
