+++ packages/EasyRepository/bundle/Enum/ConfigTag.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Bundle\Enum;
+
+enum ConfigTag: string
+{
+    case PaginatedRepository = 'easy_repository.paginated_repository';
+}
