+++ packages/EasyRepository/laravel/config/easy-repository.php
