--- packages/EasyRepository/src/Bridge/BridgeConstantsInterface.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace EonX\EasyRepository\Bridge;
-
-interface BridgeConstantsInterface
-{
-    public const TAG_PAGINATED_REPOSITORY = 'easy_repository.paginated_repository';
-}
