--- packages/EasyRepository/src/Bridge/Laravel/Exceptions/EmptyRepositoriesListException.php
@@ -1,12 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace EonX\EasyRepository\Bridge\Laravel\Exceptions;
-
-use EonX\EasyRepository\Interfaces\EasyRepositoryExceptionInterface;
-use Exception;
-
-final class EmptyRepositoriesListException extends Exception implements EasyRepositoryExceptionInterface
-{
-    // No body needed
-}
