+++ packages/EasyRepository/src/Exception/EasyPaginationNotInstalledException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Exception;
+
+use RuntimeException;
+
+final class EasyPaginationNotInstalledException extends RuntimeException implements EasyRepositoryExceptionInterface
+{
+    // No body needed
+}
