+++ packages/EasyRepository/src/Exception/EasyRepositoryExceptionInterface.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Exception;
+
+interface EasyRepositoryExceptionInterface
+{
+    // Marker for all exceptions to this package
+}
