--- packages/EasyRepository/src/Exceptions/EasyPaginationNotInstalledException.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace EonX\EasyRepository\Exceptions;
-
-final class EasyPaginationNotInstalledException extends AbstractEasyRepositoryException
-{
-    // No body needed
-}
