--- packages/EasyRepository/src/Interfaces/EasyRepositoryExceptionInterface.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace EonX\EasyRepository\Interfaces;
-
-interface EasyRepositoryExceptionInterface
-{
-    // Marker for all exceptions to this package
-}
