+++ packages/EasyRepository/src/Repository/PaginatedObjectRepositoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Repository;
+
+use EonX\EasyPagination\Interfaces\LengthAwarePaginatorInterface;
+use EonX\EasyPagination\Interfaces\PaginationInterface;
+
+interface PaginatedObjectRepositoryInterface extends ObjectRepositoryInterface
+{
+    public function paginate(?PaginationInterface $pagination = null): LengthAwarePaginatorInterface;
+}
