--- packages/EasyRepository/tests/Bridge/Laravel/Stubs/Repository2Stub.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace EonX\EasyRepository\Tests\Bridge\Laravel\Stubs;
-
-final class Repository2Stub
-{
-}
