+++ packages/EasyRepository/tests/Stub/Repository/Repository1Stub.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Tests\Stub\Repository;
+
+final class Repository1Stub
+{
+}
