+++ packages/EasyRepository/tests/Stub/Repository/Repository2Stub.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Tests\Stub\Repository;
+
+final class Repository2Stub
+{
+}
