+++ packages/EasyRepository/tests/Unit/laravel/AbstractLumenTestCase.php
@@ -1,0 +1,28 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Tests\Unit\Laravel;
+
+use EonX\EasyRepository\Tests\Unit\AbstractUnitTestCase;
+use Laravel\Lumen\Application;
+
+abstract class AbstractLumenTestCase extends AbstractUnitTestCase
+{
+    private ?Application $app = null;
+
+    protected function assertInstanceInApp(string $concrete, string $abstract): void
+    {
+        self::assertInstanceOf($concrete, $this->getApplication()->get($abstract));
+    }
+
+    protected function getApplication(): Application
+    {
+        if ($this->app !== null) {
+            return $this->app;
+        }
+
+        $this->app = new Application(__DIR__);
+
+        return $this->app;
+    }
+}
