+++ packages/EasyRepository/tests/Unit/src/Repository/DoctrineOrmRepositoryStub.php
@@ -1,0 +1,15 @@
+<?php
+declare(strict_types=1);
+
+namespace EonX\EasyRepository\Tests\Unit\Repository;
+
+use EonX\EasyRepository\Repository\AbstractDoctrineOrmRepository;
+use stdClass;
+
+final class DoctrineOrmRepositoryStub extends AbstractDoctrineOrmRepository
+{
+    protected function getEntityClass(): string
+    {
+        return stdClass::class;
+    }
+}
